pyproject.toml
[tool.poetry]
name = "minidash"
version = "0.1.0"
description = ""
authors = ["Vlad Poberezhny <vlad.pbr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
pillow = "^10.1.0"
fastapi = "^0.105.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
